using numerical index for new data structure

+++ include/field.h
@@ -681,8 +681,7 @@ struct facet {
 
 struct facet_info_t {
     // facet hash => resolved tokens
-    std::unordered_map<uint64_t, std::vector<std::string>> hashes;
-    //facet name => resolved tokens
+    std::unordered_map<uint32_t, std::vector<std::string>> hashes;
     bool use_facet_query = false;
     bool should_compute_stats = false;
     field facet_field{"", "", false};
